+++ Assets/Scripts/Characters/Units.cs
@@ -10,8 +10,18 @@ public enum UnitsTeam {
 }
 
 public enum UnitsClass {
-    OnLand,
-    Flying
+    // corps à corps
+    Tank,
+    Dps,
+
+    // milieu
+    Assassin,
+    Archer,
+    Mage,
+
+    // distance
+    Buffer,
+    Healer,
 }
 
 public enum EntityType {
@@ -41,21 +51,22 @@ public class Units : MonoBehaviour
 
     [Header("HP Bar")]
     [SerializeField] Canvas hpBarCanvas;
-    protected Slider hpSlider;
+    [SerializeField] Slider hpSlider;
     public Transform modelPosition;
 
-    [Header("Must be handle in each entities script")]
+    [Header("Must be handle in each entities script")] // voir pour un UnitData
     public float damagePerAttack = 10;
     public int manaCost = 3;
-    protected float totalHealth = 100;
-    protected float hp = 0;
-    protected float attackRate = 1f; // every seconde
-    protected float attackRange = 1.5f; // 1f is equal to 1 chess tile
-    protected float speed = 1f;
-    protected float attackAnimDuration = 1f;
+    public float totalHealth = 100;
+    public float hp = 0;
+    public float shield = 10;
+    public float attackRate = 1f; // every seconde
+    public float attackRange = 1.5f; // 1f is equal to 1 chess tile
+    public float speed = 1f;
+    public float attackAnimDuration = 1f;
     public EntityType entityType = EntityType.Basic;
-    protected float timeBeforeFirstAttack = 0f;
-    public int reward_value = 10; // Reward System
+    public int rewardValue = 10; // Reward System
+    public virtual UnitsClass unitsClass => UnitsClass.Dps;
 
 
     /* end */
@@ -72,7 +83,6 @@ public class Units : MonoBehaviour
 
     [Header("Do not touch")]
     protected string enemyTag = "Champion";
-    protected List<UnitsClass> unitsClass = new();
     protected GameObject target = null;
     protected Units targetUnitsComponent = null;
     protected GameObject[] entities = null;
@@ -89,14 +99,16 @@ public class Units : MonoBehaviour
     void Awake()
     {
         BugTracker.Info("New entity '" + gameObject.name + "' created.");
+
         animator = GetComponentInChildren<Animator>();
         currentAnimationState = AnimationState.Idle;
-        hpSlider = hpBarCanvas.GetComponentInChildren<Slider>();
         team = UnitsTeam.Enemy;
         gameObject.layer = 0;
     }
 
     protected virtual void Start() {
+        // LoadInformationsFromUnitData();
+
         isGameRunning = false;
 
         SetAnimationState(AnimationState.Idle);
@@ -107,16 +119,15 @@ public class Units : MonoBehaviour
 
         chessTileSize = VoidbornMapGeneratorHybrid.instance.chessTile;
 
-        unitsClass.Add(UnitsClass.OnLand);
-
         backupUnits.position = gameObject.transform.position;
         backupUnits.rotation = gameObject.transform.rotation;
 
+        GetComponentInChildren<ChangeHealthBarColor>().ChangeColor(team);
         if (team == UnitsTeam.Enemy)
             enemyTag = "Champion";
         else
             enemyTag = "Enemy";
-
+        
         BugTracker.Info("Entity '" + gameObject.name + "' backup created.");
     }
 
@@ -130,8 +141,18 @@ public class Units : MonoBehaviour
 
         target = null;
         Start();
-
         BugTracker.Info("Entity '" + gameObject.name + "' has been reset.");
+    }
+
+    protected void LoadInformationsFromUnitData()
+    {
+        speed = 0.5f;
+        attackRate = 2f;
+        totalHealth = 150;
+        damagePerAttack = 25;
+        manaCost = 5;
+        team = UnitsTeam.Enemy;
+        entityType = EntityType.Champion;
     }
 
     private void LateUpdate() {
@@ -216,7 +237,6 @@ public class Units : MonoBehaviour
         {
             if (entity == null || entity == gameObject)
                 continue;
-            
             Units unit = entity.GetComponentInParent<Units>();
 
             if (unit == null || !unit.isAlive)
@@ -280,7 +300,15 @@ public class Units : MonoBehaviour
         if (hpSlider == null)
             BugTracker.Error("'" + gameObject.name + "' has a hpSlider null !");
 
+        if (damage - shield < 0) {
+            damage = 0;
+            // popup shield
+            // DamagePopupManager.instance.InitShield(transform);
+        } else
+            damage -= shield;
+
         DamagePopupManager.instance.Init(transform, damage);
+
         hp -= damage;
         hpSlider.value = hp / totalHealth;
         if (hp <= 0) {
@@ -298,7 +326,7 @@ public class Units : MonoBehaviour
         target.GetComponent<Units>().TakeDamage(damagePerAttack);
         PlaySound(attackSound);
 
-        BugTracker.Info("Entity '" + gameObject.name + "' attack '"+ target.name + "' and deal '" + damagePerAttack + "' damage. (hp: "+target.GetComponent<Units>().hp+"/"+target.GetComponent<Units>().totalHealth+")");
+        BugTracker.Info("Entity '" + gameObject.name + "' attack '"+ target.name + "' and deal '" + damagePerAttack + "' damage. (hp: "+target.GetComponent<Units>().hp+"/"+target.GetComponent<Units>().totalHealth+").");
         // if (attackEffect != null)
         // {
         //     GameObject effect = Instantiate(attackEffect, target.transform.position + new Vector3(0, 10, 0), Quaternion.identity);
@@ -337,7 +365,7 @@ public class Units : MonoBehaviour
         // DONNER L'OR
         if (ShopManager.instance != null)
         {
-            ShopManager.instance.AddGold(reward_value);
+            ShopManager.instance.AddGold(rewardValue);
         }
 
         GameLoopManager.instance.CheckVictory();