+++ Assets/Scripts/Core/GameManagers.cs
@@ -1,71 +1,47 @@
+using Newtonsoft.Json;
 using UnityEngine;
 
 public class GameManager : MonoBehaviour
 {
-    public static GameManager Instance { get; private set; }
+    public static GameManager instance;
 
-    public CardUI selectedCard;
-    public ManaManager manaManager;
-    public GameObject settings;
-    public DeckData selectedDeck;
+    [Header("Current Level informations")]
+    public LevelData[] alllevelData;
+    public LevelData currentLevelData = null;
+    public string nextStory = "";
 
-    private bool isPaused = false;
+    public bool isPaused = false;
 
     private void Awake()
     {
-        Instance = this;
-    }
-
-    private void Update() {
-        if (Input.GetKeyDown(KeyCode.Escape)) {
-            if (isPaused)
-                ResumeGame();
-            else
-                PauseGame();
+        if (instance == null) {
+            instance = this;
+            DontDestroyOnLoad(gameObject);
+        } else if (instance != this) {
+            Destroy(gameObject);
         }
     }
+
+    // ---- PAUSE / RESUME ----
 
     public void PauseGame()
     {
         isPaused = true;
-        settings.SetActive(true);
         Time.timeScale = 0f;
     }
 
     public void ResumeGame()
     {
         isPaused = false;
-        settings.SetActive(false);
         Time.timeScale = 1f;
     }
 
-    public void SetSelectedCard(CardUI card)
-    {
-        if (selectedCard != null && selectedCard != card)
-        {
-            selectedCard.DeselectCard();
-        }
+    // ---- LEVELS ----
 
-        selectedCard = card;
-        Debug.Log("Carte sélectionnée : " + card.cardData.cardName);
-    }
-
-    public void DeselectCard()
+    public void SaveLevelConfig(LevelData levelData)
     {
-        if (selectedCard != null)
-        {
-            selectedCard.DeselectCard();
-            selectedCard = null;
-        }
-    }
-
-    public bool HasSelectedCard()
-    {
-        return selectedCard != null;
-    }
-
-    public GameObject GetSelectedUnitPrefab()
-    {
-        return selectedCard != null ? selectedCard.cardData.unitPrefab : null;
+        BugTracker.Info("Save current level data, current level: '" + levelData.currentlevel + "'.");
+        currentLevelData = levelData;
+        nextStory = levelData.chaptersBeforeGame;
     }
 }