+++ Assets/Scripts/SceneSpecific/Game/EnemySpawnAlgo.cs
@@ -1,25 +1,224 @@
 using System.Collections.Generic;
 using UnityEngine;
-using UnityEngine.AI;
+using System.Linq;
+using Unity.VisualScripting;
+using UnityEngine.Tilemaps;
+
+public enum Strategy {
+    SmallUnits, // plein de petites units
+    BigUnits, // peux d'units mais forte
+    Boss, // boss avec ou sans units
+    Aggressive, // max de dps
+    Turtle, // max de défence
+    Ranged // max d'attaque à distance
+}
 
 public class EnemySpawnAlgo : MonoBehaviour
 {
     public static EnemySpawnAlgo instance;
-    public int numberEnemy = 3;
+    public int manaMax = 10;
+    public int currentMana;
+    public Strategy currentStrategy;
+    public int maxEntity = 3;
+
+    private HashSet<Vector2> occupiedGridPositions = new();
+    private float _tileSize;
 
     public void Awake()
     {
         instance = this;
     }
 
+    // public void SpawnEnemies(float tileSize)
+    // {
+    //     BugTracker.Info("[Enemy Spawn Algo] Start algo spawn enemy.");
+    //     List<GameObject> allEntities = new();
+
+    //     allEntities.Add(Resources.Load<GameObject>("CrocodilePrefab"));
+    //     allEntities.Add(Resources.Load<GameObject>("TigerPrefab"));
+    //     allEntities.Add(Resources.Load<GameObject>("Enemy_tmp"));
+        
+    //     _tileSize = tileSize;
+    //     int nbrEntities = 0;
+    //     int safetyExit = 0;
+    //     currentMana = manaMax;
+
+    //     currentStrategy = Strategy.Aggressive;
+    //     BugTracker.Info("[Enemy Spawn Algo] current algo strategy: '"+currentStrategy+"'.");
+
+    //     List<GameObject> filtered = FilteredByStrategy(allEntities, currentStrategy);
+    //     List<GameObject> affordableUnits = new();
+
+    //     while (currentMana > 0 && safetyExit < 100 && nbrEntities < maxEntity) {
+    //         safetyExit++;
+
+    //         GameObject entity = filtered[Random.Range(0, filtered.Count)];
+    //         if (entity.GetComponent<Units>().manaCost - currentMana <= 0) {
+    //             currentMana -= entity.GetComponent<Units>().manaCost;
+    //             affordableUnits.Add(entity);
+    //             nbrEntities++;
+    //             BugTracker.Info("[Enemy Spawn Algo] Entity '"+entity.name+"' add to the spawn list.");
+    //         }
+    //     }
+    //     IntiateEntity(affordableUnits);
+    // }
+
+    private void IntiateEntity(List<GameObject> entities)
+    {
+        if (entities.IsUnityNull()) {
+            BugTracker.Error("[Enemy Spawn Algo] List of entities is null, failed to spawn entities.");
+            return;
+        }
+
+        foreach (GameObject e in entities) {
+            Vector2 vector2 = GetStrategicTilePosition(e.GetComponent<Units>().unitsClass);
+            Debug.Log($"x: {vector2.x} y: {vector2.y}");
+
+            GameObject entityInstance = Instantiate(e, new Vector3(vector2.x + _tileSize, 2f, vector2.y), Quaternion.identity);
+            entityInstance.transform.Rotate(0, 180, 0);
+            GameLoopManager.instance.RegisterUnit(entityInstance, false);
+
+            BugTracker.Info("'" + entityInstance.name + "' spawn.");
+        }
+    }
+
+    private Vector2 GetStrategicTilePosition(UnitsClass unitClass)
+    {
+        GameObject[] tiles = GameObject.FindGameObjectsWithTag("Tile");
+        
+        int maxX = int.MinValue;
+        int maxY = int.MinValue;
+
+        foreach (GameObject t in tiles) {
+            string[] parts = t.name.Split('_');
+            int x = int.Parse(parts[1]);
+            int y = int.Parse(parts[2]);
+
+            if (y > maxY)
+                maxY = y;
+            if (x > maxX)
+                maxX = x;
+        }
+
+        int halfBoard = maxY / 2;
+
+        int enemyZoneHeight = halfBoard + 1;
+
+        float slice = (float)enemyZoneHeight / 4f;
+
+        int RearY = 0;
+        int FrontY = halfBoard;
+
+        switch (unitClass) {
+            case UnitsClass.Healer:
+            case UnitsClass.Buffer:
+                RearY = 0;
+                FrontY = Mathf.FloorToInt((slice * 1));
+                break;
+
+            case UnitsClass.Archer:
+            case UnitsClass.Mage:
+                RearY = Mathf.FloorToInt(slice * 1);
+                FrontY = Mathf.FloorToInt((slice * 2));
+                break;
+
+            case UnitsClass.Dps:
+            case UnitsClass.Assassin:
+                RearY = Mathf.FloorToInt(slice * 2);
+                FrontY = Mathf.FloorToInt((slice * 3));
+                break;
+
+            case UnitsClass.Tank:
+                RearY = Mathf.FloorToInt(slice * 3);
+                FrontY = halfBoard;
+                break;
+        }
+
+        RearY = Mathf.Clamp(RearY, 0, halfBoard);
+        FrontY = Mathf.Clamp(FrontY, 0, halfBoard);
+
+        List<GameObject> validTiles = new();
+
+        foreach (GameObject t in tiles) {
+            string[] parts = t.name.Split('_');
+            int x = int.Parse(parts[1]);
+            int y = int.Parse(parts[2]);
+
+            if (x <= maxX && y >= RearY && y <= FrontY) {
+                if (!validTiles.Contains(t)) {
+                    validTiles.Add(t);
+                }
+            }
+        }
+
+        if (validTiles.Count == 0) {
+            foreach (GameObject t in tiles) {
+                string[] parts = t.name.Split('_');
+                int x = int.Parse(parts[1]);
+                int y = int.Parse(parts[2]);
+
+                if (x <= maxX && y >= 0 && y <= halfBoard) {
+                    if (!validTiles.Contains(t)) {
+                        validTiles.Add(t);
+                    }
+                }
+            }
+        }
+
+        if (validTiles.Count > 0) {
+            GameObject chosenTile = validTiles[Random.Range(0, validTiles.Count)];
+            occupiedGridPositions.Add(chosenTile.transform.position);
+            Debug.Log($"tile={chosenTile.name} x: {chosenTile.transform.position.x} y: {chosenTile.transform.position.y}");
+
+            return chosenTile.transform.position;
+        }
+        return new Vector2(-1f, -1f);
+    }
+
+
+    private List<GameObject> FilteredByStrategy(List<GameObject> allEntity, Strategy strategy)
+    {
+        switch (strategy) {
+            case Strategy.SmallUnits:
+                return allEntity.OrderBy(go => go.GetComponent<Units>().manaCost <= 3).ToList();
+
+            case Strategy.BigUnits:
+                return allEntity.OrderBy(go => go.GetComponent<Units>().manaCost >= 5).ToList();
+
+            // case Strategy.Boss:
+            //     return allEntity;
+
+            case Strategy.Aggressive:
+                return allEntity.Where(go => 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Dps || 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Assassin || 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Buffer).ToList();
+
+            case Strategy.Turtle:
+                return allEntity.Where(go => 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Tank || 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Healer || 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Buffer).ToList();
+
+            case Strategy.Ranged:
+                return allEntity.Where(go => 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Mage || 
+                    go.GetComponent<Units>().unitsClass == UnitsClass.Archer).ToList();
+
+            default:
+                return allEntity;
+        }
+    }
+
+
     public void SpawnEnemies(float tileSize)
     {
-        // Vector2 tilePos = GetRandomTilePosition();
-        // float x = tilePos.x + (tileSize / 2);
-        // float y = tilePos.y + (tileSize / 2);
+        Vector2 tilePos = GetRandomTilePosition();
+        // x = tilePos.x + (tileSize / 2);
+        // y = tilePos.y + (tileSize / 2);
 
         GameObject cocodilePrefab = Resources.Load<GameObject>("CrocodilePrefab");
-        // GameObject enemyInstance = Instantiate(cocodilePrefab, new Vector3(x, 2f, y), Quaternion.identity);
+        // GameObject cocodileInstance = Instantiate(cocodilePrefab, new Vector3(tilePos.x, 2f, tilePos.y), Quaternion.identity);
         GameObject cocodileInstance = Instantiate(cocodilePrefab, new Vector3(-2f, 2f, 0f), Quaternion.identity);
         cocodileInstance.transform.Rotate(0, 180, 0);
 
@@ -30,13 +229,13 @@ public class EnemySpawnAlgo : MonoBehaviour
         BugTracker.Info("CocodilePrefab spawn.");
 
         // TMP Create a Prefab
-        // while((tilePos = GetRandomTilePosition()) == tilePos);
+        while((tilePos = GetRandomTilePosition()) != tilePos);
 
         // x = tilePos.x + (tileSize / 2);
         // y = tilePos.y + (tileSize / 2);
 
         GameObject tigerPrefab = Resources.Load<GameObject>("TigerPrefab");
-        // GameObject enemyInstance = Instantiate(prefabEnemy, new Vector3(x, 2f, y), Quaternion.identity);
+        // GameObject tigerInstance = Instantiate(tigerPrefab, new Vector3(tilePos.x, 2f, tilePos.y), Quaternion.identity);
         GameObject tigerInstance = Instantiate(tigerPrefab, new Vector3(0f, 2f, 0f), Quaternion.identity);
         tigerInstance.transform.Rotate(0, 180, 0);
 
@@ -48,13 +247,13 @@ public class EnemySpawnAlgo : MonoBehaviour
         // EnemySpawnAlgo.instance.SpawnEnemies();
 
         // TMP Create a Prefab
-        // while((tilePos = GetRandomTilePosition()) == tilePos);
+        while((tilePos = GetRandomTilePosition()) != tilePos);
 
         // x = tilePos.x + (tileSize / 2);
         // y = tilePos.y + (tileSize / 2);
 
         GameObject enemy_tmpPrefab = Resources.Load<GameObject>("Enemy_tmp");
-        // GameObject enemyInstance1 = Instantiate(prefabEnemy1, new Vector3(x, 2f, y), Quaternion.identity);
+        // GameObject enemy_tmpInstance = Instantiate(enemy_tmpPrefab, new Vector3(tilePos.x, 2f, tilePos.y), Quaternion.identity);
         GameObject enemy_tmpInstance = Instantiate(enemy_tmpPrefab, new Vector3(2f, 2f, 0f), Quaternion.identity);
         enemy_tmpInstance.transform.Rotate(0, 180, 0);
 
@@ -72,8 +271,7 @@ public class EnemySpawnAlgo : MonoBehaviour
         int maxX = int.MinValue;
         int maxY = int.MinValue;
 
-        foreach (GameObject t in tiles)
-        {
+        foreach (GameObject t in tiles) {
             string[] parts = t.name.Split('_');
             int x = int.Parse(parts[1]);
             int y = int.Parse(parts[2]);
@@ -85,135 +283,13 @@ public class EnemySpawnAlgo : MonoBehaviour
         }
 
         int randomX = Random.Range(0, maxX);
-        int randomy = Random.Range(0, maxY / 2); // only half of the board
+        int randomY = Random.Range(0, maxY / 2); // only half of the board
 
         foreach (GameObject t in tiles) {
-            if (t.name == ("Tile_" + randomX + "_" + randomy))
+            if (t.name == ("Tile_" + randomX + "_" + randomY))
                 return new Vector2(t.transform.position.x, t.transform.position.y);
         }
         Debug.Log("Tile doesn't exist.");
         return new Vector2(0, 0);
     }
-
-    // public void SpawnEnemies()
-    // {
-        // GameObject[] tiles = GameObject.FindGameObjectsWithTag("Tile");
-
-        // List<GameObject> spawnTiles = new();
-
-        // int maxY = int.MinValue;
-
-        // foreach (GameObject tile in tiles)
-        // {
-        //     string[] parts = tile.name.Split('_');
-        //     int y = int.Parse(parts[2]);
-
-        //     if (y > maxY)
-        //         maxY = y;
-        // }
-
-        // foreach (GameObject tile in tiles)
-        // {
-        //     string[] parts = tile.name.Split('_');
-        //     int y = int.Parse(parts[2]);
-
-        //     if (y >= maxY - 3)
-        //         spawnTiles.Add(tile);
-        // }
-
-        // for (int i = 0; i < numberEnemy && spawnTiles.Count > 0; i++)
-        // {
-        //     int index = Random.Range(0, spawnTiles.Count);
-        //     GameObject tile = spawnTiles[index];
-        //     spawnTiles.RemoveAt(index);
-
-        //     Vector3 spawnPos = tile.transform.position + Vector3.up * 2f;
-
-        //     GameObject enemy = Instantiate(enemyPrefab, spawnPos, Quaternion.identity);
-        //     enemy.transform.Rotate(0, 180, 0);
-
-        //     enemy.GetComponent<NavMeshAgent>().enabled = false;
-
-        //     GameLoopManager.instance.RegisterUnit(enemy, false);
-        // }
-        // // TMP Create a Prefab
-        // GameObject prefabEnemy = Resources.Load<GameObject>("Enemy_tmp");
-        // GameObject enemyInstance = Instantiate(prefabEnemy, new Vector3(0f, 2f, 0f), Quaternion.identity);
-        // enemyInstance.transform.Rotate(0, 180, 0);
-
-        // // Movements are NOT managed by the navmesh
-        // enemyInstance.GetComponent<NavMeshAgent>().enabled = false;
-
-        // GameLoopManager.instance.RegisterUnit(enemyInstance, false);
-    // }
-
-
-
-
-    // public interface UnitsOnBoard
-    // {
-    //     Units units;
-    //     int tileNumber;
-    // }
-
-    // public Vector2 chessboardSize = new( 6, 6 );
-    // public Vector2 chessTileSize;
-    // public List<Units> enemyListAvailable;
-    // public List<Units> enemyListToPlaceOnBoard;
-    // public List<UnitsOnBoard> unitsOnBoards;
-    // public int DifficultyLevel = 1;
-
-    // private List board = new [
-    //     [0, 0, 0, 0, 0, 0],
-    //     [0, 0, 0, 0, 0, 0],
-    //     [0, 0, 0, 0, 0, 0],
-    //     // do not use more, this is the player board
-    //     [0, 0, 0, 0, 0, 0],
-    //     [0, 0, 0, 0, 0, 0],
-    //     [0, 0, 0, 0, 0, 0],
-    // ]; // create board form 'chessboardSize'
-
-    // private void DefineNumberOfEnemy()
-    // {
-    //     int sizeBoardEnemy = chessboardSize.Y / 2;
-    //     int nbrEnemy = sizeBoardEnemy / 4; // change this by using a tree with afinity for each units
-
-    //     for (int i = 0; i < nbrEnemy; i++) {
-    //         EnemyListToPlaceOnBoard.Append(EnemyListAvailable[Random.Next(EnemyListAvailable.Count)]);
-    //     }
-    // }
-
-    // private void PutEnemyInBoard()
-    // {
-    //     int sizeBoardEnemy = chessboardSize.Y / 2;
-
-    //     for (int i = 0; i < EnemyListToPlaceOnBoard.Count; ++i) {
-    //         bool skip = false;
-    //         int randomRow = Random.Next(chessboardSize.X);
-    //         int randomColumn = Random.Next(chessboardSize.Y / 2);
-    //         int tileNbr = randomColumn * chessboardSize.Y + randomRow;
-
-    //         for (int e = 0; unitsOnBoards.Count; e++) {
-    //             if (unitsOnBoards[i].tileNumber == tileNbr) {
-    //                 skip = true;
-    //                 i--;
-    //                 break;
-    //             }
-    //         }
-    //         if (skip)
-    //             continue;
-    //         UnitsOnBoard newUnits = { EnemyListToPlaceOnBoard[i], tileNbr };
-    //         unitsOnBoards.Add();
-    //     }
-    // }
-
-    // public void ClearEnemyListToPlaceOnBoard()
-    // {
-    //     EnemyListToPlaceOnBoard.Clear();
-    // }
-
-    // public void SelectEnemyDeck(String EnemyFolder)
-    // {
-
-    // }
 }