+++ CHANGELOG.md
@@ -2,7 +2,24 @@
 
 All notable changes to this project will be documented in this file.
 
-## [0.1.3-alpha] - 2026-
+## [0.1.4-alpha] - 2026-06-
+### Added
+- story system with json file
+- log system with UI button
+- HP bar with different color for the player and enemy
+- new level & level is now handle with a json file
+
+### Fixed
+- settings menu: sound now can be modified
+- you can not Drag and Drop units when the game is running
+- deathzone error message
+- gameManager is now a DontDestroyOnLoad & handle the whole game, menu & story
+- gameLoop now handle only the game
+- position of entity when you Drag and Drop
+- change 3D model Cavalier to humainModel
+
+
+## [0.1.3-alpha] - 2026-05-27
 ### Added
 - Drag and Drop units
 - Chess board hover color